moved qml types into models folder
