+++ src/dissolve-gui-qml.cpp
@@ -4,7 +4,7 @@
 #include "base/messenger.h"
 #include "base/processPool.h"
 #include "gui/models/dissolveModel.h"
-#include "gui/types.h"
+#include "gui/models/types.h"
 #include "main/cli.h"
 #include "main/dissolve.h"
 #include "main/version.h"
